nw eth command
