+++ cmd/Makefile
@@ -245,7 +245,7 @@ obj-$(CONFIG_ARCH_OMAP2PLUS) += ti/
 endif # !CONFIG_SPL_BUILD
 
 # Northwood
-obj-$(CONFIG_NORTHWOOD) += ./northwood
+obj-$(CONFIG_CMD_NORTHWOOD) += northwood.o
 
 obj-$(CONFIG_$(SPL_)CMD_TLV_EEPROM) += tlv_eeprom.o
